update kernel
git-svn-id: svn://svn.dd-wrt.com/DD-WRT@64409 52c4871e-980c-0410-b1e0-e73912ce01f8

+++ src/linux/universal/linux-6.6-nss/drivers/net/wireless/microchip/wilc1000/hif.c
@@ -162,7 +162,7 @@ int wilc_scan(struct wilc_vif *vif, u8 scan_source, u8 scan_type,
 	u32 index = 0;
 	u32 i, scan_timeout;
 	u8 *buffer;
-	u8 valuesize = 0;
+	u32 valuesize = 0;
 	u8 *search_ssid_vals = NULL;
 	struct host_if_drv *hif_drv = vif->hif_drv;
